+++ src/UX/ContractCreator/ViewContractModal.js
